Renamed Error.cs to ErrorMessage.cs
Also added a option for Write and WriteLine
