--- Festivity/Utils/Error.cs
@@ -1,14 +1,0 @@
-﻿using System;
-
-namespace Festivity
-{
-    class Error
-    {
-        public static void Write(string errorMessage)
-        {
-            Console.ForegroundColor = ConsoleColor.Red;
-            Console.WriteLine(errorMessage);
-            Console.ResetColor();
-        }
-    }
-}
